Raytrace almost done
Fixed method are_features_supported, moved framebuffers logic from swapchain to renderpass also fixed some other smaller issues and shaders code, still there is a bug with synchronization in raytrace manager

+++ RayTracer/Managers/Raytrace/Common/vertex.hpp
@@ -5,5 +5,5 @@ struct Vertex
 	FVector3 position;
 	FVector3 normal;
 	FVector2 uv;
-	UInt64 materialId;
+	Int32 materialId;
 };
